refactor : Used formatting tools to avoid lengthy lines and renamed few folders to maintain naming consistency
