+++ src/Components/ColorPalette/ColorPalette.jsx
