+++ src/Components/ColorPalette/colorpalette.css
