+++ src/Components/InputCard/inputcard.css
@@ -1,6 +1,6 @@
 .container-input-new-note{
     height: fit-content;
-    width: 30rem;
+    width: 40rem;
     position: relative;
     border: transparent;
     border-radius: 0.5rem;
