+++ src/Pages/Archived/archived.css
@@ -1,22 +1,7 @@
 .archived-main {
     margin-top: 6rem;
+    display: grid;
+    grid-template-columns: 1fr 2fr;
 }
 
-/* sidebar */
-.sidebar {
-    margin-left: 3rem;
-    color: var(--text-color1);
-}
-
-.sidebar-item {
-    cursor: pointer;
-    padding: 0.5rem 1.5rem;
-    width: fit-content;
-    border-radius: 1rem;
-}
-
-.sidebar-item:hover {
-    background-color: var(--secondary-color);
-    color: var(--primary-color);
-}
