+++ src/Pages/Home/Home.jsx
@@ -1,21 +1,23 @@
 import "../../css/main.css";
 import "./home.css";
-import { useState } from "react";
-import { DisplayCard, InputCard, DisplayCardEmpty, Header, Footer, Sidebar } from "../../Components";
+import { useState, useEffect } from "react";
+import { DisplayCard, InputCard, DisplayCardEmpty, Sidebar } from "../../Components";
 import { useAuth } from "../../Context";
 
 export const Home = () => {
     const [createNewCard, setCreateNewCard] = useState(false)
+    const [pinnedNotes, setPinnedNotes] = useState([])
+    const [otherNotes, setOtherNotes] = useState([])
     const { authState } = useAuth();
     const { notes } = authState;
 
-    const pinnedNotes = notes.filter((item) => item.pinned)
-    const otherNotes = notes.filter((item) => !item.pinned)
+    useEffect(() => {
+        setPinnedNotes(notes.filter((item) => item.pinned)) 
+        setOtherNotes(notes.filter((item) => !item.pinned))
+    },[notes])
 
     return(
-        <>
-        <Header />
-        <main className="home-main grid-2-column m-3">
+        <main className="home-main m-3">
             <div>
                 <Sidebar />
                 {!createNewCard && <button className="btn-create-new-note btn-solid btn-medium flex flex-justify-center flex-align-center text-base ml-3 mt-2"
@@ -53,7 +55,7 @@ export const Home = () => {
                 <div>
                     <h4 className="font-bold text-lg">Others</h4>
                     <div className="spacer-1"></div>
-                        { otherNotes.length !==0 ?
+                        {otherNotes.length !==0 ?
                             otherNotes.map((item) => {
                                 return(
                                     <DisplayCard item={item} key={item._id}/>
@@ -66,7 +68,5 @@ export const Home = () => {
             </div>
 
         </main>
-        <Footer />
-        </>
     )
 }
