+++ src/Pages/Home/home.css
@@ -1,5 +1,7 @@
 .home-main {
     margin-top: 6rem;
+    display: grid;
+    grid-template-columns: 1fr 2fr;
 }
