+++ src/css/global.css
@@ -2,6 +2,7 @@
     --primary-color : #973a99;
     --secondary-color : #faf8f8; 
     --secondary-color-light : #cfcccc1a; 
+    --secondary-color-dark : #eee5e5;  
     --tertiary-color : #b1abab;
 
     --text-color1 : #3b3939;
